refactor(client): update preact-cli
Update preact-cli to 3.0.0-rc.10 (the RC versions are being used by the
default template).

Also move `client/src/pages` to `client/src/routes` to enable route
splitting, move icons to within the `client/src` tree, and move contents
of `public` directory to the appropriate places within the `client/src`
tree.
