--- client/src/pages/index.js
@@ -1,23 +1,0 @@
-import Home from './home'
-import Login from './login'
-import Logout from './logout'
-import Profile from './profile'
-import Registration from './registration'
-import Challenges from './challenges'
-import Scoreboard from './scoreboard'
-import Error from './error'
-import Sponsors from './sponsors'
-import Verify from './verify'
-
-export {
-  Home,
-  Login,
-  Logout,
-  Profile,
-  Registration,
-  Challenges,
-  Scoreboard,
-  Error,
-  Sponsors,
-  Verify
-}
