+++ client/src/routes/logout.js
