+++ client/src/routes/profile.js
@@ -9,10 +9,10 @@ import { withToast } from '../components/toast'
 import Form from '../components/form'
 import Modal from '../components/modal'
 import util from '../util'
-import Trophy from '../../static/icons/trophy.svg'
-import AddressBook from '../../static/icons/address-book.svg'
-import UserCircle from '../../static/icons/user-circle.svg'
-import Rank from '../../static/icons/rank.svg'
+import Trophy from '../icons/trophy.svg'
+import AddressBook from '../icons/address-book.svg'
+import UserCircle from '../icons/user-circle.svg'
+import Rank from '../icons/rank.svg'
 
 const divisionMap = new Map()
