+++ client/src/routes/scoreboard.js
