+++ client/src/static/favicon.ico
