+++ package.json
@@ -7,7 +7,7 @@
     "lint": "tsc --noEmit && eslint --ext .js,.ts .",
     "start": "node dist/server/index.js",
     "migrate": "node-pg-migrate --database-url-var=RCTF_DATABASE_URL",
-    "build:client": "sh client/build.sh",
+    "build:client": "preact build --src client/src --template client/index.html --dest dist/build",
     "build:ts": "tsc && yarn copy-static",
     "build": "yarn build:ts && yarn build:client",
     "watch:client": "preact watch --src client/src --template client/index.html",
@@ -69,11 +69,11 @@
     "nodemon": "^2.0.2",
     "nyc": "^15.0.0",
     "preact": "^10.3.1",
-    "preact-cli": "^2.2.1",
+    "preact-cli": "^3.0.0-rc.10",
     "preact-router": "^3.2.1",
     "preact-svg-loader": "^0.2.1",
     "promise-polyfill": "^8.1.3",
-    "purgecss-webpack-plugin": "0.23.0",
+    "purgecss-webpack-plugin": "^2.1.0",
     "regenerator-runtime": "^0.13.3",
     "supertest": "^4.0.2",
     "typescript": "^3.8.3",
